Race condition deleting volumes
It happens that when nova server is removed, the cinder volume is
still attaching.  In this case nova won't be able to detach the
volume before removing the server.  It results in a volume attached
to a non-existing server.

The fix is to allow removing a volume if it's status is 'attached'
but the attached server no longer exists.

Closes-Bug: #1335889